+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/instructions/MultiJumpInstruction.java
@@ -15,6 +15,9 @@ public class MultiJumpInstruction extends BaseInstruction implements MultiTarget
 
     MultiJumpInstruction(AstContext astContext, List<LogicArgument> args, @Nullable List<InstructionParameterType> params) {
         super(astContext, Opcode.MULTIJUMP, args, params);
+        if (!(getArg(0) instanceof LogicAddress)) {
+            System.out.println("Ha!");
+        }
     }
 
     protected MultiJumpInstruction(BaseInstruction other, AstContext astContext) {