+++ compiler/src/main/java/info/teksol/mc/profile/options/CompilerOptionFactory.java
@@ -414,7 +414,7 @@ public class CompilerOptionFactory {
                 OptionAvailability.UNIVERSAL, category,
                 SortCategory.NONE).setValueProcessor(CompilerOptionFactory::normalizeSortVariables));
 
-        list.add(new IntegerCompilerOptionValue(DebuggingOptions.PARSE_TREE, "p",
+        list.add(new IntegerCompilerOptionValue(DebuggingOptions.PARSE_TREE, "",
                 "sets the detail level of parse tree output into the log file, 0 = off",
                 OptionMultiplicity.ONCE, SemanticStability.STABLE, OptionScope.GLOBAL,
                 OptionAvailability.UNIVERSAL, category,