+++ schemacode/src/main/java/info/teksol/schemacode/config/UnresolvedConfiguration.java
@@ -1,0 +1,5 @@
+package info.teksol.schemacode.config;
+
+public interface UnresolvedConfiguration {
+    Configuration resolve();
+}